refactor: rename internal package to source, moving LifecycleAware from API
