+++ operator-framework-core/src/main/java/io/javaoperatorsdk/operator/processing/LifecycleAware.java
@@ -1,4 +1,4 @@
-package io.javaoperatorsdk.operator.api;
+package io.javaoperatorsdk.operator.processing;
 
 import io.javaoperatorsdk.operator.OperatorException;
