--- operator-framework-core/src/main/java/io/javaoperatorsdk/operator/processing/event/internal/ResourceAction.java
@@ -1,5 +1,0 @@
-package io.javaoperatorsdk.operator.processing.event.internal;
-
-public enum ResourceAction {
-  ADDED, UPDATED, DELETED
-}
