+++ operator-framework-core/src/main/java/io/javaoperatorsdk/operator/processing/event/source/Mappers.java
@@ -1,4 +1,4 @@
-package io.javaoperatorsdk.operator.processing.event.internal;
+package io.javaoperatorsdk.operator.processing.event.source;
 
 import java.util.Collections;
 import java.util.Set;
