+++ operator-framework-core/src/main/java/io/javaoperatorsdk/operator/processing/event/source/ResourceAction.java
@@ -1,0 +1,5 @@
+package io.javaoperatorsdk.operator.processing.event.source;
+
+public enum ResourceAction {
+  ADDED, UPDATED, DELETED
+}
