+++ operator-framework-core/src/test/java/io/javaoperatorsdk/operator/processing/event/source/TimerEventSourceTest.java
@@ -1,4 +1,4 @@
-package io.javaoperatorsdk.operator.processing.event.internal;
+package io.javaoperatorsdk.operator.processing.event.source;
 
 import java.io.IOException;
 import java.util.List;
